fix(OUT-884) : download simplified, removed loader on download, created a useDownload hook
