+++ src/app/ui/NewTaskForm.tsx
@@ -36,7 +36,6 @@ import { upload } from '@vercel/blob/client'
 import { AttachmentInput } from '@/components/inputs/AttachmentInput'
 import { SupabaseActions } from '@/utils/SupabaseActions'
 import { generateRandomString } from '@/utils/generateRandomString'
-import { AttachmentCard } from '@/components/cards/AttachmentCard'
 import { bulkRemoveAttachments } from '@/utils/bulkRemoveAttachments'
 import { advancedFeatureFlag } from '@/config'
 import { WorkflowStateSelector } from '@/components/inputs/Selector-WorkflowState'
@@ -346,22 +345,6 @@ const NewTaskFormInputs = () => {
           uploadFn={(file) => uploadImageHandler(file, token ?? '', null)}
           deleteEditorAttachments={(url) => deleteEditorAttachmentsHandler(url, token ?? '', null)}
         />
-      </Stack>
-      <Stack direction="row" columnGap={2} m="16px 0px">
-        {attachments?.map((el, key) => {
-          return (
-            <Box key={el.filePath}>
-              <AttachmentCard
-                file={el}
-                deleteAttachment={async (event: React.MouseEvent<HTMLDivElement>) => {
-                  event.stopPropagation()
-                  const { data } = await supabaseActions.removeAttachment(el.filePath)
-                  store.dispatch(removeOneAttachment({ attachment: el }))
-                }}
-              />
-            </Box>
-          )
-        })}
       </Stack>
     </>
   )
